libpkg: fail on ARCHIVE_WARN
The cases here may return ARCHIVE_WARN when the necessary libfoo isn't
available, indicating that it's going to shell out to satisfy the
request.  We can't necessarily cope with shelling out, as that may be
incompatible with a sandbox.

Erroring out gives us a path to realizing that the providing libarchive
is broken.