pkg(8): Add example on how to restore databse.
As discussed recently in https://www.mail-archive.com/ports@freebsd.org/msg02505.html
add this useful example.

Using '%' as the root's prompt though not sure if this is correct. There are
three different prompts in pkg(8) manual page: $, % and #. In other FreeBSD
documentation # seems to be used for root and % for user's prompt, while here %
seems to be for root and % for normal users.