libpkg: Try to merge config files across split upgrades
libpkg will sometimes split upgrade jobs into separate delete/install
jobs in order to work around file conflicts.  At the same time, pkg
preserves config file modifications across non-split upgrades.  This
change makes pkg give consistent behaviour for both split and non-split
upgrades.

Now, split jobs are linked together with a new "xlink" field in struct
pkg_solved.  When we handle a UPGRADE_REMOVE job, the new version of the
package is passed to pkg_delete(), which uses it to check whether both
versions of the package contain a given config file; it is not removed
if so.  Then, when handling an UPGRADE_INSTALL job, we pass the
now-deleted package together with the new package, and use it to perform
a merge.

Note, this doesn't handle the case where a config file moves from one
package to another as part of an upgrade.  We should handle that too,
but the problem exists independent of whether an upgrade job is split or
not, so let's ignore it for now.  This change solves an existing problem
with pkgbase, where a split upgrade of FreeBSD-runtime can blow away
critical /etc files.

To solve the additional case described above, we will probably have to
augment the pkg_config_file structure with state which tells pkg_add()
and pkg_delete() what to do with the file.