tests: Fix a flaky repo test
The problem is that ucl_parser_add_fd() fails when parsing repo
fingerprints: it tries to lex a fingerprint as an integer and fails when
it encounters the digit "e", which causes it to use strtod().

This is not a complete fix, since pkg-repo documentation should also be
modified to include quotation marks.  Or some other solution should be
devised.