Restore support of -e with other options
Issue #2078 was based on a mis-understanding of the semantics of
the -i option. It does only apply to pkg-name pattern, not the
values inside conditional expressions.

Support for string comparisons that ignore case have been added
in a later commit, and thus there is now a way to express such
filter conditions.

This commit effectively reverts commits cacd3d7298, c40314fea, and
1bf51ad3ab7 that added the restriction of -e not being allowed in
combination with other options and which made the tests that should
have warned about the regression match the new reality.

This update should be incorporated into a pkg release as soon as
possible, since the current production version of pkg broke the
conflicts checks in the ports system ("malke check-copnflicts").