curl: reimplement mirror_type=http
This version only implements the second version (aka the simple one)
and not the original which to my knowledge have never existed.

As a reminder this is how it works

pkg will fetch the list of mirror in the form of the text content
formatted like follow:
URL: <mirror1url>
URL: <mirror2url>

this is the result of an http get on the url provided as repo url.
changes vs the previous implementation, the text get is now limited
 to 1MB