libpkg: don't complain about Last-Modified on failed fetches
We avoid complaining for HTTP 304, but let's also not complain if we've
failed to fetch entirely (either out of retries or simply cancelled) to
reduce the noise a little bit when packages simply haven't been built
yet.  More importantly, it avoids leading one to believe that it's
perhaps a misconfigured header setting rather than simply missing.