utils.c: return NULL when getenv(...) fails in get_http_auth() function
In get_http_auth() function, whenever getenv(...) fails it returns
a boolean, "false" which itself is 0. Compiler automatically casts
this, as (char *)0 (equivalent to NULL), and never emits a warning.

This works, including (return 0;, which also gets treated as NULL)
but this is technically wrong. Explicitly specifying the symbol,
NULL, seems a better choice.

Signed-off-by: rilysh <nightquick@proton.me>