libpkg/pkg_elf.c: avoid calling strlen() function to re-calculate the length
There are several places where strlen() is used to re-calculate
the length. Although possibly Clang and GCC _may_ convert some
of them to pure functions, it's not for every strlen() call.

Also, both Clang and GCC ignores strlen() in a loop, and will
call the function as many times as long as the loop will be.

strlen() test (calculating length in the loop)
https://godbolt.org/z/3vzh69Ynq

strlen() test (already calculated outside of the loop)
https://godbolt.org/z/YvqWe4WT3

Signed-off-by: rilysh <nightquick@proton.me>