Universal support for both big and little endian architectures
The conversion of the ELF hints file to little endian format has been
proposed in https://reviews.freebsd.org/D44053 (rtld and ldconfig -
this review will be split to allow the changes to rtld to be approved
before ldconfig is modified to generally write little-endian files).

As an intermediate step, both the ldconfig command and run time loader
ld-elf.so.1 shall support either byte-order.

The ldconfig command shall create little-endian hints files by
default, with a new option -B to support the creation of big-endian
hints files (e.g. for FreeBSD version with big-endian only ld-elf.so.1
on powerpc64).

In preparation for the use of little-endian hints files on big-endian
architectures (effectively only powerpc64), the pkg command should be
made compatible with eiher byte-order.

Since this is a pre-requisite for the commit to ldconfig that makes it
create little-endian hints files by default, the pkg-devel port should
be enhanced to support either byte order as early as possible.

Actually, the non-devel version of pkg will also need to be enhanced
in this way, before ldconfig can create little-endian hints files on
powerpc64.