pkg: add a `key --sign` operation to sign arbitrary data
This pulls arbitray data from stdin, then runs it through the usual signing
machinery as-is with no further digesting or processing than what the
underlying signer does.  For OpenSSL, this means that we sign the SHA256 of
the raw data, and the output is the raw signature.  For the ECC signer,
we'll either sign the SHA512 or SHA256 depending on whether it's eddsa or
ecdsa.

This was useful for verifying the signature output of the pubkey mechanism
when debugging, and I don't see any harm in exposing it -- the signature
written can be verified with OpenSSL for both ECDSA and the OSSL signer.