docs: update documentation to account for the ECC changes
The main user-facing changes are that pkg-repo needs to know the signer to
use for the key if it's not the rsa (ossl) signer, and the signing server
needs to output a TYPE if it's not an rsa signature.  The client will pick
up the correct signer type to use from metadata that pkg adds to signatures
and/or public keys included in the archive.