src/info.c: keep the length in a variable to avoid another call
Keeping the string length (strlen()) value in a variable can avoid
re-calculating the length of the string (or another strlen() function
call). Although, most compilers might just "fix" this by keeping it
in the previous register, but it will only, if optimization is enabled.
To avoid this, it seems just better to do this manually.

Signed-off-by: rilysh <nightquick@proton.me>