libpkg/pkg_repo.c: multiple trivial changes
1. In pkg_repo_archive_extract_check_archive() function, don't set rc
value that needed to be returned whenever an error or okay happens.
Instead, just return the error to avoid a few more instructions to
set the rc variable and jumping to a different label.

2. pkg_repo_fetch_data_fd(): do the check, whether fd is -1 in the same
braces where the function was called. This avoids two multiple checks
for fd if the first one isn't -1.

e.g. if fd isn't -1
and below another check which will be always true if the previous one
succeed.
if fd isn't -1
...

3. Use the environment variable "TMPDIR" and not "TMMDIR". (It doesn't
exists) and so the branch will always going to be false.

4. Same as 2.

5. Same as 4.

Signed-off-by: rilysh <nightquick@proton.me>