libpkg: push siglen +/- 1 into the ossl signer
pack_sign() cannot know that it's safe to read past the bounds of the buffer
provided, so we have to stop that.  We can't just stop writing the nul
terminator to avoid breaking compatibility with previous versions' signing
and verifying RSA signatures, so instead push the hack into the ossl signer.

This fixes an overflow when trying to read from the libder-encoded buffer.