libpkg/../query.c: use constness, test branch once, and others
* These SQL strings aren't being modified later and generally, they're
just string constants. Use const to explicitly mark them as constants.

* Avoid initializing bsql value with basesql, as if match is MATCH_INTER-
NAL, bsql value have to be set again (previously the value was basesql),
which will use a mov instruction. Use a branch (if..else) to remove that.

* "how" variable is initialized just below, and the function will return
NULL, when none of the case matches with the provided one. So the initia-
lization to NULL is unneeded here.

Signed-off-by: rilysh <nightquick@proton.me>