extraction: fix an upgrade issue with symlinks becoming directories
many users took the habit to create symlinks from random directories
into another partition when running out of disk space, for this reason
when pkg extracts files it does accept to extract files below a symlink
if this symlink is valid and points to a directory.

The problem is if a package itself provide a symlink to a directory and
then a latter upgrade is providing a directory instead of that symlink.
it results and unsolveable extraction problem for pkg.

From now on, pkg will not accept anymore symlinks not owned by pkg to
exists when extracting files, with the exception of prefix which is
according to the report the number 1 case of symlinks done directly by
users.