backup.c: do not overwrite the "d" pointer
In pkgdb_dump(), whenever xstrdup() function is called, it's return
value is stored in the d variable. On the next line get_dirname()
is called and the return value of it is also stored in the same d
variable. Since strdup() does allocate memory, so the previous d
pointer for xstrdup() will get lost, resulting a memory leak.

Signed-off-by: rilysh <nightquick@proton.me>