Improve `pkgdb_access` API
`pkgdb` works with multirepo, however, it's API is not aware that we
sometimes want to work with some specific repo and not with all of them

The idea of this PR is to allow all `pkgdb_*` functions to be able
to select a set of repos to work with (e.g. that could be overriden
from CLI args).

The only option for now is to disable repos in the configuration which
might be not convenient for some use cases.