libpkg: implement new job scheduler
The old job scheduling implementation had multiple bugs and lacked a
robust formal model, making the code hard to understand and fix.

Most importantly, it does not provide hard guarantees about job
ordering and leaves the ordering under-specified (See PR 259785).

The new scheduler is designed from first principles and models the
problem as a directed graph where jobs are nodes and directed edges
represent scheduling order requirements.

The existence of a cycle in the graph indicates that an upgrade job
needs to be split. Once all cycles have been removed, a topological sort
on the graph gives the execution order.

In the absence of strict ordering requirements the new scheduler orders
execution in a way that minimizes the distance between the split halves
of a split upgrade job in the execution order, a feature that the old
scheduler was not capable of.

PR:		259785
Sponsored by:	The FreeBSD Foundation