tests/conflicts: update for new job scheduler
In both the complex_conflicts and more_complex_choice test cases.
the new stricter job scheduler splits the upgrade foo job.

The reasoning is that the currently installed version of foo depends on
the currently installed version of bar so it would be invalid to delete
bar while the currently installed version of foo is still installed.

This creates a cycle in the job scheduling graph which is solved by
splitting the upgrade foo job.

Sponsored by:	The FreeBSD Foundation