tests/conflicts-multirepo: update for new job scheduler
The old job scheduler unnecessarily split the upgrade of test2 in the
conflicts_multirepo test case. The new scheduler does not split this
upgrade, so update the test case to match.

Additionally, add a modified version of the test case where it is
necessary to split the upgrade to test that path as well.

Sponsored by:	The FreeBSD Foundation