tests/issue1425: update for new job scheduler
The new scheduler sorts jobs lexicographically in the absence of hard
ordering requirements.

There's a bit more going on in this test case though, pkgA indirectly
depends on pkgD and it could be argued that pkgD should therefore be
reinstalled first.

However, I would argue that there is no need for a hard ordering
requirement in this case. Regardless of the order, all dependencies of
both pkgA and pkgD are installed at the point in time the job is
executed. Therefore, it should not be possible to observe any difference
between the two possible orderings from a PRE-INSTALL or POST-INSTALL
script.

Sponsored by:	The FreeBSD Foundation