Fix kyua tests/lua:print_msg on Ubuntu 24.04 and MacOS.
lua:print_msg unit test fails on recent Ubuntu/MacOS failing to compare the testfile (Cannot open testfile).
The issue is with the inappropriate call to open(O_CREATE) which mandates a third mode argument. If unspecified the mode of the file is a random number from the stack and might be lacking read permission for the subsequent comparison to succeed.

While here added ATF checks for open/close.