Fix host component handling in file:// URLs
pkg handles file:// URLs in libpkg/fetch_file.c which translates the file URL into an access to the local filesystem.
Ignoring the host component altogether in this process is semantically wrong.
The host component, if given, directs the file to be fetched from that host, not from the local filesystem.

Extend the handling of the host component to accept the localhost host component - the null host component file:/// is anyway accepted already, but reject any other host component, as the code is not intended to go around looking for files on other hosts.