Introduce --with-tsan to compile with Clang ThreadSanitizer
pkg repo uses multiple threads to speed up the repository creation process. On MacOS with multiple workers this is predictably crashing ```pkg``` or generating unparsable repository artefacts.
Introduce --with-tsan to compile with Clang ThreadSanitizer.

While here, make the description of the configure options more clear and remove ```-ggdb``` in favour of ```-g```.