Consistently use Darwin and other cosmetics
pkg uses the OS, which under MacOS is Darwin. Rename PKG_OS_MACOS to PKG_OS_DARWIN to be consistent in the source as well.

Cater for a corner case in having a basename in Mach-O dynamic library.

Adjust comments and simplify parsing code in pkg_abi_macho after the major abi refactoring.