pkg-repo: use buffered stdio to write output files
Before the switch to threaded repo creation, file locks and unbuffered
writes were used to prevent interleaved output.  Now that it's all threaded
and the critical section is protected by a mutex, remove the remaining flock
calls and switch to stdio.  Removes thousands of write() syscalls and saves
~8 seconds of CPU time on my system.

There is no fopenat() function, so the openat() calls are left in place but
fdopen() is called on the results and the FILE* it returns is used after
that.