libpkg: fall back to Linux for ELF shlib parsing
If the target ABI is either Linux or FreeBSD and the target OS of
a shared library ELF file cannot be determined, assume that it is a
Linux ELF file.

It is unfortunately impossible to reliably determine whether a shared
library targets Linux.

See the new comment in pkg_elf.c for further details.

Sponsored by:	The FreeBSD Foundation