extract: fix regression when dealing with symlinks in the path
e482b6668b813359ad3658ad5c5c1432964e7b1b introduced a regression:
this commit to not accept directories other that prefix to be a symlink
to another location, if the symlink itself was not owned by pkg.

The last part which is checking that the symlink was owned by a package
was the broken one which is fixed here. now pkg will fallback in
following symlinks if it could find a package which is not itself which
owns the said symlinks.