update: drastically reduce memory usage
with the new repo the data metadata is now a valid "big" json (54MB)
Parsing and converting it to a ucl object takes around 270MB of memory
Alternate json parsers tested (yyjson, cjson etc) takes roughly the same
amount of memory except for jsmn parser which takes roughly 100MB for
the same json.

Use this parser when dealing with repositories, which makes pkg updates
usable again on system with low memory (tested of VM with 512MB and no
swap).