pkg add: cache the span of the filesystem
Only perform once the loop over all the packages to gather the
provides/requires to be used when calling pkg add command.

This was attempted in pkg 1.20 already but backout due to performance
penalty, by caching to once per call of pkg add this should improve
the performance penalty.

Next step if this is not enough would be to save the cache for other
calls