abi: accept symlinks in valid path to be analysed
the elf analysis now accept to follow symlinks, the but differs to
after the full analysis to decide if the shlib provided should be
exposed by the package or not. this way we can check of the shlib
provided we should is really exposed by a regular file which is not
in the valid path but actually owned by the package as we don't want
to exposed libraries which are owned by other packages.

While here when keeping the internal_provided libraries, keep them with
their decorator, so where doing the cleanup later we are sure we are
cleaning all the internal_provided library and not only the native one
which can happen with multilib packages