libpkg: Provide the METALOG path to pkg scripts
This is useful for non-root package installations.  For instance, the
pkg post-install and post-deinstall for ca_root_nss just runs "certctl
rehash", but certctl can take a METALOG path parameter.  With this
feature and a bit of work, that post-install script can be run as a
non-root user.

Sponsored by:	The FreeBSD Foundation
Sponsored by:	Klara, Inc.