pkg_delete: Run script even for split upgrade
Even if the package will be reinstalled after we want to run the *DEINSTALL
scripts as it can cause problem for example for @sample file.
If a package update the @sample and it's marked as a split upgrade the script
needs to be run otherwise the @sample file will never be updated.

Sponsored by:	Beckhoff Automation GmbH & Co. KG