scheduler: simplify topological sort
There is some accidental complexity here after the universal move from
linked lists to vectors in pkg. Really what this code wants is a set
type, but an unordered vector is fine as well.

Add a vec_swap_remove() function to aid unordered usage and simplify the
topological sort implementation. In theory, this could be a small
performance win as well due to getting rid of a branch in the loop of
pkg_jobs_schedule_has_incoming_edge().

Sponsored by:	The FreeBSD Foundation