libpkg: fix resuming of transfers with libcurl
The fetcher will be passed an offset as part of its fetch_item() if it
expects to resume a transfer, which is our cue to pass that along to
libcurl.  Pass it along to CURLOPT_RESUME_FROM_LARGE to avoid messing
up a large interrupted transfer by downloading the entire file into the
partial, then having to download the entire thing again because the
checksum clearly wouldn't match.

A number of places have to be modified to understand that HTTP 206 is
an acceptable response in lieu of HTTP 200 in case we are trying to
resume the transfer.