pkgdb: track uname, gname, permissions and fflags in pkgdb
and allow pkg-query to list those attributes using new format
specifiers.

The idea is to embed the file and directory metadata (user, group,
permissions, file flags) during package creation time in +MANIFEST.
Later, this metadata will be written into the package database
during install time. Support for querying metadata is provided in this
patch.

This patch does not change the way which metadata is used during
installation time. The metadata used during package installation
still relies on the tar embedded metadata (which should be equal
to the metadata in the +MANIFEST, but we don't verify this yet).

Partially addresses #1088 and allows to implement a metadata aware
pkg-check later.