libpkg: Add a heuristic for backup_libraries in shlib handling
When BACKUP_LIBRARIES is enabled and a package containing shlibs is
upgraded, we can assume that old versions of those libraries are
preserved.  For example, when libutil.so was bumped from 9 to 10, an
attempt to upgrade the pkgbase base system (correctly) ended up removing
all of my ports packages since they depend on libutil.so.9.  However,
when BACKUP_LIBRARIES is configured, we have fewer constraints;
following the upgrade, libutil.so.9 will still be available.

Hack the solver so that it recognizes this situation.  We could instead
synthesize the compat-libraries package which contains backed-up
libraries earlier, and add it to the job universe such that the solver
automatically incorporates it, but that approach is more complex and
doesn't have any obvious benefits versus this one.

See also #2501.