libpkg: Relax ABI checking when looking for shlib providers
Suppose one is upgrading a pkgbase system from FreeBSD:15:amd64 to
FreeBSD:16:amd64.  The ports packages will all have shlib dependencies
on libc.so.7, belonging to some pkgbase package.  Previously, when
upgrading just the base system, the solver would not allow the FreeBSD
16 libc to satisfy the dependency since it has a different ABI.
However, this doesn't make sense: libc.so, and shared libraries in
general, provide backward compatibility.  A newer OS version in the ABI
shouldn't prevent them from satisfying a dependency.

Thus, add a heuristic to pkg_solve_handle_provide() to relax the ABI
checking for this case: if the provider of the shlib has an ABI with a
matching OS and arch, let it satisfy the dependency so long as its OS
version isn't older than that of the requiring package.

In general packages from a given repo will all have a matching ABI so
this should only affect cross-repo shlib dependencies.

See also #2520.