deps: make sure the origin is only an information
Remove exact patch between both name and origin, this is not necessary
anymore since we move the unicity very long ago to the name. besides
origin can be changed on the fly and we want to still be able to get the
exact dependency chain when doing pkg info -d.