libpkg: Add METALOG entries for intermediate directories
When installing a package, pkg might need to create a directory not
listed in the manifest.  In that case, it wasn't adding METALOG entries.
For instance, if a package installs the single file /foo/bar/baz/config,
pkg will create /foo, /foo/bar, and /foo/bar/baz implicitly, and these
should be logged.

Add a metalog_add() call to open_tempdir() to effect this.  Basically,
when preparing to install a file or directory at a given path,
open_tempdir() looks for the longest prefix of the path which
corresponds to an existing directory.  As it descends, we know that each
visited path doesn't exist and so will be created later, so we can emit
a metalog entry there.

There are other ways to implement this without a lot of refactoring, but
they either miss some cases or result in many duplicate METALOG entries.

Fixes: #2531