periodic/400.status-pkg: Add -C and -R to pkg version
-C prevents false positives such as:

FreeBSD-ftpd-15.0.a2.20250919175449 needs updating (remote has 20250929)

(the 'updated' package is "freebsd-ftpd".)

-R causes pkg to avoid consulting /usr/ports to determine out-of-date
packages, which breaks with pkgbase because base packages are not in
ports:

FreeBSD-yp-15.0.a5.20251010043546 orphaned: base/FreeBSD-yp

Although this is a functional change, most users probably want it to
consult the package database, not the local ports tree.  Checking for
outdated ports is really outside the purview of pkg(8).