libpkg: Ignore file conflicts when registering a backup library
The BACKUP_LIBRARIES may try to back up the same shlib multiple times.
This on its own is mostly harmless, but it results in a conflict when
re-registering the compat-libraries package, and pkgdb_register_pkg()
fails.  If this happens during a package deletion caused by a split
upgrade, the backuplib transaction savepoint is not running within a
nested transaction, and a subsequent attempt to start a transaction will
result in an error, since rolling back to a savepoint doesn't cancel the
transaction.

Fix the problem by simply force-installing the package.

Fixes: #2522