scheduler: loosen ordering for upgrade jobs
Currently the scheduler splits pretty much every upgrade job in
practice. This is because when both the old and new packages for A
depend on the old and new packages for B respectively, the strict rules
of the scheduler see a cycle and don't allow upgrading B while A is
install to avoid a state where old A is installed but old B is not.

I believe this rule to be too strict when applied to upgrade jobs and
have loosened the "old depends on old" rule to ignore upgrade jobs where
the "new depends on new" rule also applies. This patch passes all
existing test cases, including the test cases checking that split
upgrades are done when necessary. Therefore, I do not think the new
rules are too loose.

This patch also adds a basic test case that was not passed with the old
logic, I'm very surprised we didn't have a test checking this.

Sponsored by:	The FreeBSD Foundation