ABI: fix abi globs validation
in the solver when we do install check for shlibs provides, we do verify
 if the ABI string from the package we might install is compatible with
the ABI of the package which requires the provide.

While we theorically expect exactly ABI given we are speaking of shlibs,
it has been quite common in the past to accept and use wildcards
(starting with compat-libraries package).

Relax ABI verification to accept wildcard everywhere, to avoid spitting
error messages like "Unknown OS '*' in ABI string".

Extra note: this commit might be reverted in a couple of years.