libpkg: improve curl fetch error reporting
Currently all errors that occur in curl_do_fetch() are reported to the
user as "Failed to fetch X: No Error" which is both incorrect and very
unhelpful.

With this patch, the correct error message from libcurl is reported
to the user, for example "Failed to fetch X: Could not resolve hostname"

This patch does not fix all the error handling done by the libcurl
fetching code, but it does fix the case that seems to be most commonly
hit by users.

Closes:		https://github.com/freebsd/pkg/issues/2498
Sponsored by:	The FreeBSD Foundation