Add unregister
The goal of unregister is to remove the record from the database that
a package is install but to no touch the files (or even run the scripts
as we want to keep the generated data for example).
This will allow people to migrate from a pkgbase to a non-pkgbase install.

Sponsored by:	Beckhoff Automation GmbH & Co. KG